+++ quimb/tensor/fitting.py
@@ -4,7 +4,6 @@ from autoray import compose, dag, do
 
 from ..utils import check_opt
 from .contraction import contract_strategy
-from .array_ops import isfermionic
 
 
 def tensor_network_distance(
@@ -95,7 +94,7 @@ def tensor_network_distance(
     if method == "dense":
         tnA = tnA.contract(output_inds=oix, preserve_tensor=True)
         tnB = tnB.contract(output_inds=oix, preserve_tensor=True)
-        if isfermionic(tnA.data):
+        if tnA.isfermionic():
             # if fermion tensor, flip dual outer indices in A
             data = tnA.data
             dual_outer_axs = tuple(